+++ crates/faucet/src/lib.rs
@@ -114,14 +114,26 @@ impl Faucet {
         let sqlite_store = SqliteStore::new(config.store_path.clone()).await?;
 
         let mut client = ClientBuilder::new()
-            .tonic_rpc_client(&config.node_endpoint, Some(config.timeout.as_millis() as u64))
+            .grpc_client(&config.node_endpoint, Some(config.timeout.as_millis() as u64))
             .authenticator(Arc::new(keystore))
             .store(Arc::new(sqlite_store))
             .build()
             .await?;
+
+        // We sync to the chain tip before importing the account to avoid matching too many notes
+        // tags from the genesis block (in case this is a fresh store).
+        client
+            .sync_state()
+            .instrument(info_span!(target: COMPONENT, "faucet.load.sync_state"))
+            .await
+            .context("faucet failed to sync state")
+            .inspect_err(|err| {
+                error!(?err, "failed to sync state");
+            })?;
+        client.ensure_genesis_in_place().await?;
+
         client.add_account(&account, false).await?;
         client.set_setting(DEFAULT_ACCOUNT_ID_SETTING.to_owned(), account.id()).await?;
-        client.ensure_genesis_in_place().await?;
 
         if deploy {
             let mut faucet = Self::load(config).await?;
@@ -139,7 +151,7 @@ impl Faucet {
     pub async fn load(config: &FaucetConfig) -> anyhow::Result<Self> {
         let span = tracing::Span::current();
         let client = ClientBuilder::new()
-            .tonic_rpc_client(&config.node_endpoint, Some(config.timeout.as_millis() as u64))
+            .grpc_client(&config.node_endpoint, Some(config.timeout.as_millis() as u64))
             .filesystem_keystore(KEYSTORE_PATH)
             .store(Arc::new(SqliteStore::new(config.store_path.clone()).await?))
             .build()
@@ -210,6 +222,18 @@ impl Faucet {
         &mut self,
         requests: impl IntoIterator<Item = (MintRequest, MintResponseSender)>,
     ) -> anyhow::Result<()> {
+        // We sync before creating the transaction to ensure the state is up to date. If the
+        // previous transaction somehow failed to be included in the block, our state would
+        // be out of sync.
+        self.client
+            .sync_state()
+            .instrument(info_span!(target: COMPONENT, "faucet.mint.sync_state"))
+            .await
+            .context("faucet failed to sync state")
+            .inspect_err(|err| {
+                error!(?err, "failed to sync state");
+            })?;
+
         let span = tracing::Span::current();
 
         let (valid_requests, response_senders) = self.filter_requests_by_supply(requests);
@@ -238,15 +262,6 @@ impl Faucet {
         span.record("tx_id", tx_id.to_string());
 
         Self::send_responses(response_senders, note_ids, tx_id);
-
-        self.client
-            .sync_state()
-            .instrument(info_span!(target: COMPONENT, "faucet.mint.sync_state"))
-            .await
-            .context("faucet failed to sync state")
-            .inspect_err(|err| {
-                error!(?err, "failed to sync state");
-            })?;
         Ok(())
     }
 
@@ -426,4 +441,118 @@ fn build_p2id_notes(
         notes.push(note);
     }
     Ok(notes)
+}
+
+#[cfg(test)]
+mod tests {
+    use std::env::temp_dir;
+
+    use miden_client::ExecutionOptions;
+    use miden_client::account::component::AuthRpoFalcon512;
+    use miden_client::account::{AccountBuilder, AccountStorageMode, AccountType};
+    use miden_client::asset::TokenSymbol;
+    use miden_client::auth::AuthSecretKey;
+    use miden_client::crypto::rpo_falcon512::SecretKey;
+    use miden_client::store::{NoteFilter, Store};
+    use miden_client::testing::MockChain;
+    use miden_client::testing::mock::{MockClient, MockRpcApi};
+    use miden_client_sqlite_store::SqliteStore;
+    use tokio::sync::{mpsc, oneshot};
+
+    use super::*;
+    use crate::types::NoteType;
+
+    #[tokio::test]
+    async fn batch_requests() {
+        let batch_size = 32;
+
+        let (tx_mint_requests, rx_mint_requests) = mpsc::channel(1000);
+        let mut receivers = vec![];
+        for i in 0..batch_size {
+            let (sender, receiver) = oneshot::channel();
+            let mint_request = MintRequest {
+                account_id: AccountId::try_from(1).unwrap(),
+                note_type: if i % 2 == 0 {
+                    NoteType::Public
+                } else {
+                    NoteType::Private
+                },
+                asset_amount: AssetAmount::new(100_000_000).unwrap(),
+            };
+            tx_mint_requests.send((mint_request, sender)).await.unwrap();
+            receivers.push(receiver);
+        }
+
+        let store =
+            Arc::new(SqliteStore::new(temp_dir().join("batch_requests.sqlite3")).await.unwrap());
+        run_faucet(rx_mint_requests, batch_size, store.clone());
+
+        for receiver in receivers {
+            let response = receiver.await.unwrap().unwrap();
+            let notes = store.get_output_notes(NoteFilter::Unique(response.note_id)).await.unwrap();
+            assert_eq!(notes.len(), 1);
+        }
+    }
+
+    // TESTING HELPERS
+    // ---------------------------------------------------------------------------------------------
+
+    /// Runs a faucet on a separate thread using a mock client.
+    fn run_faucet(
+        rx_mint_requests: mpsc::Receiver<(MintRequest, MintResponseSender)>,
+        batch_size: usize,
+        store: Arc<dyn Store>,
+    ) {
+        let secret = SecretKey::new();
+        let symbol = TokenSymbol::new("TEST").unwrap();
+        let max_supply = Felt::try_from(1_000_000_000_000_u64).unwrap();
+        let account = AccountBuilder::new(rand::random())
+            .account_type(AccountType::FungibleFaucet)
+            .storage_mode(AccountStorageMode::Public)
+            .with_component(BasicFungibleFaucet::new(symbol, 6, max_supply).unwrap())
+            .with_auth_component(AuthRpoFalcon512::new(secret.public_key()))
+            .build()
+            .unwrap();
+        let key = AuthSecretKey::RpoFalcon512(secret);
+
+        std::thread::spawn(move || {
+            // Create a new runtime for this thread
+            let rt = tokio::runtime::Builder::new_current_thread()
+                .enable_all()
+                .enable_io()
+                .build()
+                .expect("Failed to build runtime");
+
+            // Run the faucet on this thread's runtime
+            rt.block_on(async {
+                let keystore =
+                    FilesystemKeyStore::<StdRng>::new(PathBuf::from("keystore")).unwrap();
+                keystore.add_key(&key).unwrap();
+
+                let mut client = MockClient::new(
+                    Arc::new(MockRpcApi::new(MockChain::new())),
+                    Box::new(RpoRandomCoin::new(Word::empty())),
+                    store,
+                    Some(Arc::new(keystore)),
+                    ExecutionOptions::new(None, 4096, false, false).unwrap(),
+                    None,
+                    None,
+                    None,
+                )
+                .await
+                .unwrap();
+                client.ensure_genesis_in_place().await.unwrap();
+                client.add_account(&account, false).await.unwrap();
+                let faucet = Faucet {
+                    id: FaucetId::new(account.id(), NetworkId::Testnet),
+                    client,
+                    tx_prover: Arc::new(LocalTransactionProver::default()),
+                    issuance: Arc::new(RwLock::new(AssetAmount::new(0).unwrap())),
+                    max_supply: AssetAmount::new(1_000_000_000_000).unwrap(),
+                    script: TransactionScript::read_from_bytes(TX_SCRIPT).unwrap(),
+                };
+                faucet.run(rx_mint_requests, batch_size).await.unwrap();
+            });
+        });
+    }
 }